Initial react-blessed support
