+++ CHANGELOG
@@ -1,3 +1,6 @@
+0.9.2
+* React Blessed support. import Yellowbox from 'yellowbox-react/lib/blessed';
+
 0.9.1
 * Don't publish .babelrc to npm
